- Move chunk splitting to PerformanceEstimator subclasses - Move roc-auc based min_chunk_size predictor to CBPE (also ROC-AUC based).
